Change multiplier variables from float to double

+++ src/main/java/io/github/webexistence/classicenchanting/config/ClassicEnchantingConfig.java
@@ -8,8 +8,8 @@ import eu.midnightdust.lib.config.MidnightConfig;
 public class ClassicEnchantingConfig extends MidnightConfig {
     public static final String GENERAL = "general";
 
-    @Entry(category = GENERAL, isSlider = true, min=0.0f, max=2.5f, precision = 4)
-    public static float enchantmentLevelCostMultiplier = 1.0F;
+    @Entry(category = GENERAL, isSlider = true, min=0.0f, max=2.5, precision = 4)
+    public static double enchantmentLevelCostMultiplier = 1.0;
     @Entry(category = GENERAL)
     public static boolean enableGrindstoneMultiplier = true;
 }
